Validate Timestamp conversion from SystemTimes before UTC Epoch time
The protobuf spec for `Timestamp` states that:

    negative second values with fractions must still have
    non-negative nanos values that count forward in time

... but in v0.6.1 and earlier of `prost_types`, this is handled wrongly
when converting to `SystemTime`, with the `nanos` value contributing to
the `Duration` that is to be _subtracted_ from the `UNIX_EPOCH`:
https://github.com/danburkert/prost/blob/v0.6.1/prost-types/src/lib.rs#L142-L145

```Rust
let mut duration = Duration {
    seconds: -timestamp.seconds,
    nanos: timestamp.nanos,
};
```

So, for example `Timestamp { seconds: -20, nanos: 999_999_999 }` would
result in a `Duration { seconds: 20, nanos: 999_999_999 }` to subtract
from `UNIX_EPOCH`, instead of the correct value of 19 seconds + 1 nano
before `UNIX_EPOCH`.

The bug was fixed in commit 4ad5e3451fcd2e2192406351671a962e2e2be909
(https://github.com/danburkert/prost/pull/297), which replaces the old
`TryFrom` implementation with a full `From<Timestamp> for SystemTime`,
and adds `From<SystemTime> for Timestamp` support for pre-UTC Epoch time
`SystemTime` values.  However, the existing unittests do not include any
specific validation for timestamps earlier than the UTC epoch time.

This patch therefore adds a set of representative test cases which make
sure that the protobuf spec is respected.  This should prevent the risk
of regressions in future.  These are provided in two separate functions:

  * `check_timestamp_negative_seconds` provides general purpose
    checks, with nanos values provided in 100 ns increments so
    as to guarantee the same behaviour on all platforms (POSIX
    `SystemTime` has 1 ns resolution, but Windows only 100 ns)

  * `check_timestamp_negative_seconds_1ns` provides tests that
    include nanos values down to 1 ns resolution, which will
    only work on POSIX systems

Part of https://github.com/danburkert/prost/issues/327.